Test demonstration of issue
When the window is resized to be smaller than the Image inside the
ImageView, the ImageView stops scaling to match the size of the window.
This is the issue, since, in the UpdateWindowImpl class, background
images are usually either extremely high quality/resolution or are
simply just larger than the window. I thought that the background
ImageViews' parent would resize the ImageViews and shrink them to fit,
but instead, the ImageViews just prevent their parent from shrinking,
and stay the same size themselves.

The workaround is to load the ImageViews' images in with sizes that are
just under or exactly at the predicted size that will be allocated for
the ImageViews' parent.

See the comment, in UpdateWindowImpl, that says:

// ~~ ImageView issue workaround ~~

for the workaround itself.