Fixed clicking in resize margin
If the user clicks in the resize margin or drags the mouse too short of
a distance, the ResizeOperator will synthesize and fire a ClickEvent at
the position where the user released the short drag/click.

This makes it so that the ResizeOperator does not consume simple clicks
and other things when the user intended to click and not resize
something. This was an issue with the close button, whereas clicking in
the top right quarter (or so) of the close button would simply do
nothing.